Split more targets from chrome.gyp to add support for building Native Client for 64-bit Windows.
This CL is required for supporting Native Client on 64-bit Windows. On Win64 we will use a 64-bit executable to load NaCl modules. Since we have only one version of Chrome on Windows, this executable has to be built together with 32-bit Chrome binaries.
For the targets that are required for both 32-bit and 64-bit builds, this CL moves the shared parts of the target definitions into target-specific gypi files (such as base.gypi). New 64-bit only targets are added. Since we would like to port as few code as possible at this point, some 64-bit targets include less source files or less dependencies than the original 32-bit targets. These targets will have nacl_win64 in their name - to emphasize that they should be used for NaCl only. Some 64-bit targets are identical to the 32-bit targets in everything except for the target platform - these will have only win64 in their name. 

TEST=none
BUG=28176
Review URL: http://codereview.chromium.org/384068

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@34983 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3ec54a2fe0482dc97113478ac87e45b505e636be

+++ webkit/glue/websocketstreamhandle_delegate.h
@@ -21,14 +21,14 @@ class WebSocketStreamHandleDelegate {
   virtual void WillOpenStream(WebKit::WebSocketStreamHandle* handle,
                               const GURL& url) {}
   virtual void WillSendData(WebKit::WebSocketStreamHandle* handle,
-                            const char* data, int len) {}
+                            const char* data, size_t len) {}
 
   virtual void DidOpenStream(WebKit::WebSocketStreamHandle* handle,
                              int max_amount_send_allowed) {}
   virtual void DidSendData(WebKit::WebSocketStreamHandle* handle,
                            int amount_sent) {}
   virtual void DidReceiveData(WebKit::WebSocketStreamHandle* handle,
-                              const char* data, int len) {}
+                              const char* data, size_t len) {}
   virtual void DidClose(WebKit::WebSocketStreamHandle*) {}
 };
